+++ AlchemistCode/Cards/Rare/BottomsUp.cs
@@ -1,0 +1,25 @@
+using Alchemist.AlchemistCode.Powers;
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.Models.Powers;
+
+namespace Alchemist.AlchemistCode.Cards.Rare;
+
+public class BottomsUp : AlchemistCard
+{
+    public override CardMultiplayerConstraint MultiplayerConstraint =>
+        CardMultiplayerConstraint.MultiplayerOnly;
+
+    public BottomsUp() : base(2, CardType.Power, CardRarity.Rare, TargetType.Self)
+    {
+        WithCostUpgradeBy(-1);
+        WithTip(typeof(PoisonPower));
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await PowerCmd.Apply<BottomsUpPower>(choiceContext, Owner.Creature, 1, Owner.Creature, this);
+    }
+}